Hacking this app to be a document ranking app.
It will currently load the list of people, then load document rankings for a selected person.

The ranking columns need to take up more width.  The ranking columns need descriptions and better titles.  The autocomplete is too slow.  I need to add person and document information displays.  The entire thing is a hack and needs some TLC and refactoring to be better.