Show user rankings.
Adjust to latest IST database.

Let a user say if an entity is a possible match and what their confidence is.  Show this in the lineup.

Refactored a bunch of js to make it pass jshint and jscsrc tests.

Note that there are some hacks to work around some IST data issues.