Better data set specification.
Allow multiple data sets to be specified independently for the left and right sides.  Allow a pretty name to be given for a dataset so we don't have to show the collection name if we don't want to.  If there is only a single data set on a side, don't show a selection control.  Automatically select the first data set on each side.

This changes the defaults.json file.  Instead of 'twitter' and 'instagram', there are two new keys: 'graph1Datasets' and 'graph2Datasets'.  Each of this is a LIST.  Each entry in the list may either be a string (in which case the value is both the collection in Mongo AND the display text) or an object with two keys: 'key' is the Mongo collection name, and 'name' is the display text.  The order of the list is the same order that will be shown to the user.  The first entry will be automatically loaded.

For instance, as a fragment of the defaults.json:

```
"graph1Datasets": [{
  "key": "twitter_isil_36_nodelink", "name": "Twitter ISIL 36"
}, {
  "key": "twitter_isil_36", "name": "Some other Twitter"
}],
"graph2Datasets": ["twitter_isil_36_nodelink"],
```

On the left, there are two entries, which the user will see as "Twitter ISIL 36" and "Some other Twitter".  "twitter_isil_36_nodelink" will be automatically loaded.  On the right, the user will see "twitter_isil_36_nodelink" as the sole value and not be able to change it.