Reset node neighborhood query limits.
Each time a node is expanded, we query additional neighbors.  This allows handling large neighbors by limiting the amount that is shown by default.  However, each time a node was queries, the limit would be increased if the node had been visible in the previous graph.  By reseting it, selecting an entity will always result in the same 1-hop network prior to expanding nodes.